chore: Surfaced the folders in src
