--- src/migration/index.ts
@@ -1,1 +1,0 @@
-export * from './ember-addon/index.js';
