+++ src/steps/analyze-addon.ts
