+++ src/steps/move-addon-files.ts
