+++ src/steps/move-project-root-files.ts
