+++ src/steps/update-addon-package-json.ts
