+++ src/steps/update-addon-package-json/index.ts
