+++ src/steps/update-addon-package-json/update-dependencies.ts
