+++ src/steps/update-addon-package-json/update-dev-dependencies.ts
