+++ src/steps/update-addon-package-json/update-other-fields.ts
