+++ src/steps/update-addon-package-json/update-scripts.ts
