+++ src/steps/update-addon-tsconfig-json.ts
