+++ src/steps/update-test-app-package-json/move-dev-dependencies-to-dependencies.ts
