+++ src/steps/update-test-app-package-json/update-dependencies.ts
