+++ src/steps/update-test-app-package-json/update-other-fields.ts
